+++ src/main/kotlin/com/booleworks/kjobs/api/persistence/hashmap/HashMapPersistence.kt
@@ -60,9 +60,29 @@ open class HashMapJobPersistence : JobPersistence, JobTransactionalPersistence {
         PersistenceAccessResult.result(jobs.values.filter {
             it.status in setOf(
                 JobStatus.SUCCESS,
-                JobStatus.FAILURE
+                JobStatus.FAILURE,
+                JobStatus.CANCELLED
             ) && it.finishedAt?.isBefore(date) ?: false
         }.map { it.copy() })
+
+    override suspend fun allJobsExceedingDbJobCount(maxNumberKeptJobs: Int): PersistenceAccessResult<List<Job>> {
+        val exceedingJobCount = jobs.count() - maxNumberKeptJobs
+        return if (exceedingJobCount > 0) {
+            jobs.values.filter {
+                it.status in setOf(
+                    JobStatus.SUCCESS,
+                    JobStatus.FAILURE,
+                    JobStatus.CANCELLED
+                )
+            }.let { finishedJobs ->
+                PersistenceAccessResult.result(finishedJobs.sortedBy { job -> job.createdAt }.take(exceedingJobCount).map { it.copy() })
+            }
+        } else {
+            PersistenceAccessResult.result(emptyList())
+        }
+
+    }
+
 
     override suspend fun persistJob(job: Job): PersistenceAccessResult<Unit> = PersistenceAccessResult.success.also { jobs[job.uuid] = job }