+++ src/main/kotlin/com/booleworks/kjobs/api/persistence/redis/RedisPersistence.kt
@@ -15,6 +15,7 @@ import com.booleworks.kjobs.data.PersistenceAccessResult
 import com.booleworks.kjobs.data.internalError
 import com.booleworks.kjobs.data.modified
 import com.booleworks.kjobs.data.notFound
+import com.booleworks.kjobs.data.orQuitWith
 import com.booleworks.kjobs.data.result
 import com.booleworks.kjobs.data.success
 import com.booleworks.kjobs.data.uuidNotFound
@@ -32,6 +33,7 @@ import kotlinx.coroutines.withContext
 import org.slf4j.LoggerFactory
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
+
 
 private val logger = LoggerFactory.getLogger("com.booleworks.kjobs.RedisDataPersistence")
 
@@ -84,7 +86,8 @@ open class RedisJobPersistence(
             RedisJobTransactionalPersistence(commands, config).run { block() }
             val transactionResult = commands.exec().get()
             if (transactionResult == null || transactionResult.wasDiscarded()) {
-                logger.debug("Transaction with precondition was not executed ({})",
+                logger.debug(
+                    "Transaction with precondition was not executed ({})",
                     transactionResult?.let { "Transaction was discarded" } ?: "Transaction was null")
                 PersistenceAccessResult.modified()
             } else {
@@ -125,9 +128,26 @@ open class RedisJobPersistence(
 
     override suspend fun allJobsFinishedBefore(date: LocalDateTime): PersistenceAccessResult<List<Job>> =
         getAllJobsBy({ commands, key -> commands.hmget(key, "status", "finishedAt") }) {
-            it.get("status") in setOf(JobStatus.SUCCESS.toString(), JobStatus.FAILURE.toString())
+            it.get("status") in setOf(JobStatus.SUCCESS.toString(), JobStatus.FAILURE.toString(), JobStatus.CANCELLED.toString())
                     && it.get("finishedAt")?.let(LocalDateTime::parse)?.isBefore(date) ?: false
         }
+
+    override suspend fun allJobsExceedingDbJobCount(maxNumberKeptJobs: Int): PersistenceAccessResult<List<Job>> {
+        val overallKeyCount = scanKeys(config.jobPattern).count()
+        val exceedingJobCount = overallKeyCount - maxNumberKeptJobs
+        return if (exceedingJobCount > 0) {
+            getAllJobsBy({ commands, key -> commands.hmget(key, "status") }) {
+                it.get("status") in setOf(JobStatus.SUCCESS.toString(), JobStatus.FAILURE.toString(), JobStatus.CANCELLED.toString())
+            }.orQuitWith {
+                logger.error("Failed fetching jobs for job count: " + it.message)
+                return PersistenceAccessResult.internalError<List<Job>>(it.message)
+            }.let { jobs ->
+                PersistenceAccessResult.result(jobs.sortedBy { it.createdAt }.take(exceedingJobCount))
+            }
+        } else {
+            PersistenceAccessResult.result(emptyList())
+        }
+    }
 
     override suspend fun fetchStates(uuids: List<String>): PersistenceAccessResult<List<JobStatus>> {
         // TODO use pipelining